Move logic.py from eudlib to ctrlstru
