+++ eudplib/collections/eudarray.py
@@ -11,6 +11,7 @@ from typing_extensions import Self
 from .. import core as c
 from .. import utils as ut
 from ..core.inplacecw import iand, ilshift, ior, irshift, iset, isub, ixor
+from ..ctrlstru import EUDNot
 from ..localize import _
 from ..memio import f_dwread_epd, f_dwwrite_epd
 
@@ -220,8 +221,6 @@ class EUDArray(ut.ExprProxy):
         return c.MemoryEPD(self._epd + key, c.Exactly, val)
 
     def neitem(self, key, val):
-        from ..eudlib.utilf import EUDNot
-
         self._bound_check(key)
         return EUDNot(c.MemoryEPD(self._epd + key, c.Exactly, val))
 
@@ -234,13 +233,9 @@ class EUDArray(ut.ExprProxy):
         return c.MemoryEPD(self._epd + key, c.AtLeast, val)
 
     def ltitem(self, key, val):
-        from ..eudlib.utilf import EUDNot
-
         self._bound_check(key)
         return EUDNot(c.MemoryEPD(self._epd + key, c.AtLeast, val))
 
     def gtitem(self, key, val):
-        from ..eudlib.utilf import EUDNot
-
         self._bound_check(key)
         return EUDNot(c.MemoryEPD(self._epd + key, c.AtMost, val))
