+++ eudplib/ctrlstru/logic.py
@@ -5,10 +5,10 @@
 # and is released under "MIT License Agreement". Please see the LICENSE
 # file that should have been included as part of this package.
 
-from eudplib import core as c
-from eudplib import ctrlstru as cs
-from eudplib.trigger.tpatcher import negate_cond
-from eudplib.utils.eperror import EPError
+from .. import core as c
+from ..trigger.tpatcher import negate_cond
+from ..utils.eperror import EPError
+from .simpleblock import EUDElse, EUDElseIf, EUDElseIfNot, EUDEndIf, EUDIf, EUDIfNot
 
 
 def EUDOr(cond1, *conds) -> c.EUDLightBool:  # noqa: N802
@@ -20,14 +20,14 @@ def EUDOr(cond1, *conds) -> c.EUDLightBool:  # noqa: N802
     """
 
     v = c.EUDLightBool()
-    if cs.EUDIf()(cond1):
+    if EUDIf()(cond1):
         c.RawTrigger(actions=v.Set())
     for cond in conds:
-        if cs.EUDElseIf()(cond):
+        if EUDElseIf()(cond):
             c.RawTrigger(actions=v.Set())
-    if cs.EUDElse()():
+    if EUDElse()():
         c.RawTrigger(actions=v.Clear())
-    cs.EUDEndIf()
+    EUDEndIf()
     return v
 
 
@@ -45,14 +45,14 @@ def EUDAnd(cond1, *conds) -> c.EUDLightBool:  # noqa: N802
     """
 
     v = c.EUDLightBool()
-    if cs.EUDIfNot()(cond1):
+    if EUDIfNot()(cond1):
         c.RawTrigger(actions=v.Clear())
     for cond in conds:
-        if cs.EUDElseIfNot()(cond):
+        if EUDElseIfNot()(cond):
             c.RawTrigger(actions=v.Clear())
-    if cs.EUDElse()():
+    if EUDElse()():
         c.RawTrigger(actions=v.Set())
-    cs.EUDEndIf()
+    EUDEndIf()
     return v
 
 
@@ -79,7 +79,7 @@ def EUDNot(cond):  # noqa: N802
         pass
     v = c.EUDLightBool()
     c.RawTrigger(actions=v.Set())
-    if cs.EUDIf()(cond):
+    if EUDIf()(cond):
         c.RawTrigger(actions=v.Clear())
-    cs.EUDEndIf()
+    EUDEndIf()
     return v
